requirements-dev.txt
-r base.txt     # include base.txt

django-debug-toolbar==3.1.1
flake8==3.8.4
mixer==6.1.3
pycodestyle==2.6.0
pyflakes==2.2.0